Split NotificationManager - add NotificationStatus and RegisterDevice
